refactor(entities): Break Entity commands out to specific files
